hive-dj-bot/requirements.txt
aws-cdk-lib==2.171.1
constructs>=10.0.0,<11.0.0
python-dotenv==1.0.1
